# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Orca API
  version: 1.1.0
  description: |-
    Multi-Agent Orchestration System API

    Manage agents, sessions, analytics, billing, webhooks, alerts, backups, and more.
  contact:
    name: Elkhedr
    url: https://github.com/ekagent/elkhedr-orca
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Register user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  format: password
                role:
                  type: string
                  enum:
                    - user
                    - manager
                    - admin
                    - guest
                  default: user
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/login:
    post:
      operationId: loginUser
      summary: Login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usernameOrEmail
                - password
              properties:
                usernameOrEmail:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/orca/actions:
    get:
      operationId: listOrcaActionRequests
      summary: List Orca bridge action requests
      tags:
        - Orca
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orca action request list
    post:
      operationId: createOrcaActionRequest
      summary: Create an approval-gated Orca bridge action request
      tags:
        - Orca
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Orca action request created
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/orca/actions/{actionId}:
    get:
      operationId: getOrcaActionRequest
      summary: Read one Orca bridge action request
      tags:
        - Orca
      security:
        - bearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orca action request
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orca/actions/{actionId}/approval:
    post:
      operationId: decideOrcaActionRequest
      summary: Approve or reject a pending Orca bridge action request
      tags:
        - Orca
      security:
        - bearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orca action decision recorded
        "409":
          description: Action request is not pending approval
  /api/orca/actions/{actionId}/result:
    post:
      operationId: completeOrcaActionRequest
      summary: Attach a result to an approved Orca bridge action request
      tags:
        - Orca
      security:
        - bearerAuth: []
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orca action result recorded
        "409":
          description: Action request has not been approved
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - manager
            - user
            - guest
        created_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - invalid or missing parameters
    Unauthorized:
      description: Unauthorized - valid authentication required
...
    NotFound:
      description: Resource not found